fix00_part1

+++ libft00/ft_pow.c
@@ -6,11 +6,11 @@
 /*   By: anonymous <anonymous@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/04 14:12:07 by anonymous         #+#    #+#             */
-/*   Updated: 2017/11/04 14:15:09 by anonymous        ###   ########.fr       */
+/*   Updated: 2017/11/06 23:15:37 by achernys         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int     ft_pow(int num, int pow)
+long	ft_pow(long num, int pow)
 {
     if (pow == 0)
         return (1);
